refactor(kwin-effect): extract filter helpers to plasmazoneseffect/window_filtering.cpp
Pulls 12 filter/sticky/active-window helpers out of plasmazoneseffect.cpp:
ensurePreSnapGeometryStored, buildWindowMap, getValidActiveWindowOrFail,
isWindowFloating, shouldHandleWindow, isTileableWindow,
hasOtherWindowOfClassWithDifferentPid, isDaemonReady,
syncFloatingWindowsFromDaemon, getActiveWindow, isWindowSticky,
updateWindowStickyState. Drops the main file from 5021 to 4779 lines.
