+++ kwin-effect/CMakeLists.txt
@@ -43,6 +43,7 @@ set(plasmazones_effect_SRCS
     plasmazoneseffect/borders.cpp
     plasmazoneseffect/window_identity.cpp
     plasmazoneseffect/mouse_drag.cpp
+    plasmazoneseffect/window_filtering.cpp
     autotilehandler.cpp
     autotilehandler.h
     autotilehandler/signals.cpp
