+++ kwin-effect/plasmazoneseffect.cpp
@@ -348,80 +348,6 @@ inline GLenum wrapStringToEnum(const QString& wrap)
 
 // NavigateDirectivePrefix moved to navigationhandler.cpp to avoid redefinition
 
-// ═══════════════════════════════════════════════════════════════════════════════
-// D-Bus helpers (all async — no QDBusInterface to avoid synchronous introspection)
-// ═══════════════════════════════════════════════════════════════════════════════
-
-// ═══════════════════════════════════════════════════════════════════════════════
-// Helper Method Implementations
-// ═══════════════════════════════════════════════════════════════════════════════
-
-void PlasmaZonesEffect::ensurePreSnapGeometryStored(KWin::EffectWindow* w, const QString& windowId,
-                                                    const QRectF& preCapturedGeometry)
-{
-    if (!w || windowId.isEmpty()) {
-        return;
-    }
-
-    if (!isDaemonReady("ensure pre-snap geometry")) {
-        return;
-    }
-
-    // Use pre-captured geometry if provided, otherwise read from window.
-    QRectF geom = preCapturedGeometry.isValid() ? preCapturedGeometry : w->frameGeometry();
-    if (geom.width() <= 0 || geom.height() <= 0) {
-        return;
-    }
-
-    // Use virtual-screen-aware ID — getWindowScreenId() falls back to the physical
-    // ID when virtual screen defs haven't loaded yet, so it is safe to call
-    // unconditionally. Using it here ensures the stored screen ID always matches
-    // the ID used by later lookups.
-    const QString screenId = getWindowScreenId(w);
-
-    // Post the store directly with overwrite=false. The daemon's storePreTileGeometry
-    // enforces per-windowId idempotency — a second capture for the same runtime
-    // instance is a no-op. We deliberately skip the prior async hasPreTileGeometry
-    // pre-check: that path matched on appId too, so a stale cross-session entry from
-    // a prior window instance (keyed by appId) would block the fresh per-instance
-    // capture and freeze float-restore at ancient coordinates.
-    PhosphorProtocol::ClientHelpers::fireAndForget(
-        this, PhosphorProtocol::Service::Interface::WindowTracking, QStringLiteral("storePreTileGeometry"),
-        {windowId, static_cast<int>(geom.x()), static_cast<int>(geom.y()), static_cast<int>(geom.width()),
-         static_cast<int>(geom.height()), screenId, false},
-        QStringLiteral("storePreTileGeometry"));
-    qCInfo(lcEffect) << "Stored pre-tile geometry for window" << windowId << "geom=" << geom;
-}
-
-QHash<QString, KWin::EffectWindow*> PlasmaZonesEffect::buildWindowMap(bool filterHandleable) const
-{
-    const auto windows = KWin::effects->stackingOrder();
-    QHash<QString, KWin::EffectWindow*> windowMap;
-    windowMap.reserve(windows.size());
-    for (KWin::EffectWindow* w : windows) {
-        if (w && (!filterHandleable || shouldHandleWindow(w))) {
-            windowMap[getWindowId(w)] = w;
-        }
-    }
-    return windowMap;
-}
-
-KWin::EffectWindow* PlasmaZonesEffect::getValidActiveWindowOrFail(const QString& action)
-{
-    KWin::EffectWindow* activeWindow = getActiveWindow();
-    if (!activeWindow || !shouldHandleWindow(activeWindow)) {
-        qCDebug(lcEffect) << "No valid active window for" << action;
-        emitNavigationFeedback(false, action, QStringLiteral("no_window"));
-        return nullptr;
-    }
-    return activeWindow;
-}
-
-bool PlasmaZonesEffect::isWindowFloating(const QString& windowId) const
-{
-    return m_navigationHandler->isWindowFloating(windowId);
-}
-
 bool PlasmaZonesEffect::supported()
 {
     // This effect is a compositor plugin that works in KWin on Wayland
@@ -1324,130 +1250,6 @@ void PlasmaZonesEffect::slotSettingsChanged()
     // emits windowsTiled directly for those changes.
 }
 
-bool PlasmaZonesEffect::shouldHandleWindow(KWin::EffectWindow* w) const
-{
-    if (!w) {
-        return false;
-    }
-
-    // Never snap our own overlay/editor windows (but allow the settings app)
-    const QString windowClass = w->windowClass();
-    if (windowClass.contains(QLatin1String("plasmazonesd"), Qt::CaseInsensitive)
-        || windowClass.contains(QLatin1String("plasmazones-editor"), Qt::CaseInsensitive)) {
-        return false;
-    }
-
-    // Exclude XDG desktop portal windows (file dialogs, color pickers, etc.)
-    if (windowClass.contains(QLatin1String("xdg-desktop-portal"), Qt::CaseInsensitive)) {
-        return false;
-    }
-
-    // Plasma shell layer-shell surfaces — see isPlasmaShellSurface() for rationale.
-    if (isPlasmaShellSurface(windowClass)) {
-        return false;
-    }
-
-    // Check user-configured exclusion lists (needed for drag gating — daemon also enforces
-    // for keyboard nav, but the effect must filter for drag operations and lifecycle reporting)
-    if (!m_excludedApplications.isEmpty() || !m_excludedWindowClasses.isEmpty()) {
-        KWin::Window* kw = w->window();
-        const QString appName = kw ? kw->desktopFileName() : QString();
-        for (const QString& excluded : m_excludedApplications) {
-            if (!excluded.isEmpty() && appName.contains(excluded, Qt::CaseInsensitive)) {
-                return false;
-            }
-        }
-        for (const QString& excluded : m_excludedWindowClasses) {
-            if (!excluded.isEmpty() && windowClass.contains(excluded, Qt::CaseInsensitive)) {
-                return false;
-            }
-        }
-    }
-
-    // Skip special / non-manageable window types (inherently effect-side — KWin metadata)
-    if (w->isSpecialWindow() || w->isDesktop() || w->isDock() || w->isFullScreen() || w->isSkipSwitcher()) {
-        return false;
-    }
-
-    // Skip transient/dialog windows unconditionally. Dialogs, utilities, tooltips,
-    // notifications, etc. should never be zone-managed. User-configured exclusion
-    // lists and minimum size checks are handled by the daemon.
-    if (w->isDialog() || w->isUtility() || w->isSplash() || w->isNotification() || w->isOnScreenDisplay()
-        || w->isModal() || w->isPopupWindow()) {
-        return false;
-    }
-
-    return true;
-}
-
-bool PlasmaZonesEffect::isTileableWindow(KWin::EffectWindow* w) const
-{
-    // Reject menus, popups, tooltips, modals, and transient children.
-    // Electron apps (Vesktop, VS Code, Discord) create separate windows
-    // for context menus and dropdowns that pass shouldHandleWindow() but
-    // must never enter the autotile tree.
-    if (!w->isNormalWindow() || w->isModal() || w->isPopupWindow() || w->isDropdownMenu() || w->isPopupMenu()
-        || w->isTooltip() || w->isMenu() || w->transientFor()) {
-        return false;
-    }
-    // Reject keep-above windows — overlay/utility tools (Spectacle, color
-    // pickers, screen rulers, etc.) set keep-above and should not enter the
-    // autotile tree or receive auto-focus. Without this guard, opening
-    // Spectacle while focusNewWindows is enabled disrupts the tiled layout.
-    if (w->keepAbove()) {
-        return false;
-    }
-    return true;
-}
-
-// shouldAutoSnapWindow removed — equivalent to shouldHandleWindow + isTileableWindow.
-// Call sites use isTileableWindow directly (stricter than shouldHandleWindow alone).
-
-bool PlasmaZonesEffect::hasOtherWindowOfClassWithDifferentPid(KWin::EffectWindow* w) const
-{
-    if (!w) {
-        return false;
-    }
-
-    QString windowClass = w->windowClass();
-    pid_t windowPid = w->pid();
-
-    // Check all existing windows for same class but different PID
-    // This detects when another app (e.g., Cachy Update) spawns a window
-    // of a class that the user has previously snapped (e.g., Ghostty)
-    const auto windows = KWin::effects->stackingOrder();
-    for (KWin::EffectWindow* other : windows) {
-        if (other == w) {
-            continue; // Skip self
-        }
-        if (!shouldHandleWindow(other)) {
-            continue; // Skip non-managed windows
-        }
-        if (other->windowClass() == windowClass && other->pid() != windowPid) {
-            // Found another window of the same class with different PID
-            // This means the new window was likely spawned by a different app
-            return true;
-        }
-    }
-
-    return false;
-}
-
-bool PlasmaZonesEffect::isDaemonReady(const char* methodName) const
-{
-    if (!m_daemonServiceRegistered) {
-        qCDebug(lcEffect) << "Cannot" << methodName << "- daemon not ready";
-        return false;
-    }
-    return true;
-}
-
-void PlasmaZonesEffect::syncFloatingWindowsFromDaemon()
-{
-    // Delegate to NavigationHandler
-    m_navigationHandler->syncFloatingWindowsFromDaemon();
-}
-
 // Template implementation for loadSettingAsync — delegates to shared helper.
 template<typename Fn>
 void PlasmaZonesEffect::loadSettingAsync(const QString& name, Fn&& onValue)
@@ -1727,26 +1529,6 @@ void PlasmaZonesEffect::connectNavigationSignals()
                                           SLOT(slotSnapAssistReady(QString, QString, PhosphorProtocol::EmptyZoneList)));
 
     qCInfo(lcEffect) << "Connected to navigation D-Bus signals";
-}
-
-KWin::EffectWindow* PlasmaZonesEffect::getActiveWindow() const
-{
-    // Prefer KWin's active (focused) window when it is manageable and on current desktop
-    KWin::EffectWindow* active = KWin::effects->activeWindow();
-    if (active && active->isOnCurrentActivity() && active->isOnCurrentDesktop() && !active->isMinimized()
-        && shouldHandleWindow(active)) {
-        return active;
-    }
-    // Fallback: topmost manageable window on current desktop (e.g. when activeWindow() is
-    // null or refers to a dialog/utility we don't handle)
-    const auto windows = KWin::effects->stackingOrder();
-    for (auto it = windows.rbegin(); it != windows.rend(); ++it) {
-        KWin::EffectWindow* w = *it;
-        if (w && w->isOnCurrentActivity() && w->isOnCurrentDesktop() && !w->isMinimized() && shouldHandleWindow(w)) {
-            return w;
-        }
-    }
-    return nullptr;
 }
 
 QString PlasmaZonesEffect::outputScreenId(const KWin::LogicalOutput* output) const
@@ -2766,32 +2548,6 @@ void PlasmaZonesEffect::callResolveWindowRestore(KWin::EffectWindow* window, std
 // The kwin-effect no longer calls the legacy dragStarted D-Bus method;
 // beginDrag sets up snap-path state internally on the daemon side, so
 // there's only one code path into the drag state machine.
-bool PlasmaZonesEffect::isWindowSticky(KWin::EffectWindow* w) const
-{
-    return w && w->isOnAllDesktops();
-}
-
-void PlasmaZonesEffect::updateWindowStickyState(KWin::EffectWindow* w)
-{
-    if (!w || !m_daemonServiceRegistered) {
-        return;
-    }
-
-    QString windowId = getWindowId(w);
-    if (windowId.isEmpty()) {
-        return;
-    }
-
-    bool sticky = isWindowSticky(w);
-    // Use fire-and-forget instead of QDBusInterface to avoid synchronous D-Bus
-    // introspection. slotWindowAdded → updateWindowStickyState fires for every
-    // window during login; QDBusInterface creation blocks the compositor thread
-    // for ~25s if the daemon hasn't entered app.exec() yet (daemonReady is
-    // emitted before the event loop starts).
-    PhosphorProtocol::ClientHelpers::fireAndForget(this, PhosphorProtocol::Service::Interface::WindowTracking,
-                                                   QStringLiteral("setWindowSticky"), {windowId, sticky},
-                                                   QStringLiteral("setWindowSticky"));
-}
 
 // The dragMoved lambda sends updateDragCursor directly via
 // ClientHelpers::fireAndForget. Single entry point for hot-path cursor updates.
